requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ipware==7.0.1
pillow==11.2.1
python-ipware==3.0.0
sqlparse==0.5.3
